+++ classes/WpMatomo/Admin/GeolocationSettings.php
@@ -41,7 +41,7 @@ class GeolocationSettings implements AdminSettingsInterface {
 			 && check_admin_referer( self::NONCE_NAME )
 			 && current_user_can( Capabilities::KEY_SUPERUSER ) ) {
 
-			$maxmind_license = stripslashes($_POST[ self::FORM_NAME ]);
+			$maxmind_license = trim(stripslashes($_POST[ self::FORM_NAME ]));
 
 			if (empty($maxmind_license)) {
 				$maxmind_license = '';