+++ classes/WpMatomo/Report/Renderer.php
@@ -16,9 +16,45 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 class Renderer {
+	CONST CUSTOM_UNIQUE_ID_VISITS_OVER_TIME = 'visits_over_time';
 
 	public function register_hooks() {
 		add_shortcode( 'matomo_report', array( $this, 'show_report' ) );
+	}
+
+	public function show_visits_over_time($limit)
+	{
+		$cannot_view = $this->check_cannot_view();
+		if ($cannot_view) {
+			return $cannot_view;
+		}
+
+		if (is_numeric($limit)) {
+		    $limit = (int) $limit;
+        } else {
+            $limit = 14;
+        }
+
+		$report_meta = array('module' => 'VisitsSummary', 'action' => 'get');
+
+		$data = new Data();
+		$report = $data->fetch_report($report_meta, 'day', 'last' . $limit, 'label', $limit);
+		$first_metric_name = 'nb_visits';
+
+		ob_start();
+
+		include 'views/table_map_no_dimension.php';
+
+		return ob_get_clean();
+	}
+
+	private function check_cannot_view()
+	{
+		if ( ! current_user_can( Capabilities::KEY_VIEW ) ) {
+			// not needed as processRequest checks permission anyway but it's faster this way and double ensures to not
+			// letting users view it when they have no access.
+			return esc_html__( 'Sorry, you are not allowed to view this report.', 'matomo' );
+		}
 	}
 
 	public function show_report( $atts ) {
@@ -31,10 +67,17 @@ class Renderer {
 			$atts
 		);
 
-		if ( ! current_user_can( Capabilities::KEY_VIEW ) ) {
-			// not needed as processRequest checks permission anyway but it's faster this way and double ensures to not
-			// letting users view it when they have no access.
-			return esc_html__( 'Sorry, you are not allowed to view this report.', 'matomo' );
+		$cannot_view = $this->check_cannot_view();
+		if ($cannot_view) {
+			return $cannot_view;
+		}
+
+		if ($a['unique_id'] === 'visits_over_time') {
+		    $is_default_limit = $a['limit'] === 10;
+		    if ($is_default_limit) {
+		        $a['limit'] = 14;
+            }
+			return $this->show_visits_over_time($a['limit']);
 		}
 
 		$metadata    = new Metadata();