+++ tests/phpunit/wpmatomo/user/test-sync.php
@@ -88,19 +88,33 @@ class UserSyncTest extends MatomoAnalytics_TestCase {
 
 		$idsite = $this->get_current_site_id();
 
+		switch_to_blog($blogid1);
+		$user2 = get_user_by('login', 'admin');
+		restore_current_blog();
+		switch_to_blog($blogid2);
+		$user3 = get_user_by('login', 'admin');
+		restore_current_blog();
+
 		$this->assertCount( 1, $this->mock->synced_users[0]['users'] );
 		unset( $this->mock->synced_users[0]['users'] );
+
+        $this->assertCount( 1, $this->mock->synced_users[1]['users'] );
+        $this->assertEquals( $user2->ID, $this->mock->synced_users[1]['users'][0]->ID );
+        unset( $this->mock->synced_users[1]['users'] );
+
+        $this->assertCount( 1, $this->mock->synced_users[2]['users'] );
+        $this->assertEquals( $user3->ID, $this->mock->synced_users[2]['users'][0]->ID );
+        unset( $this->mock->synced_users[2]['users'] );
+
 		$this->assertEquals(
 			array(
 				array(
 					'idSite' => $idsite,
 				),
 				array(
-					'users'  => array(),
 					'idSite' => $idsite,
 				),
 				array(
-					'users'  => array(),
 					'idSite' => $idsite,
 				),
 			),